Move some classes to matching packets
