+++ adapter/src/main/java/eu/wswieciejutra/controller/ViewController.java
@@ -1,5 +1,7 @@
-package eu.wswieciejutra;
+package eu.wswieciejutra.controller;
 
+import eu.wswieciejutra.Facade;
+import eu.wswieciejutra.LoggerUtility;
 import eu.wswieciejutra.dto.BranchDto;
 import eu.wswieciejutra.dto.RepositoryDto;
 import eu.wswieciejutra.exception.UserNotFoundException;
