+++ adapter/src/main/java/eu/wswieciejutra/persistence/SqlBranchAdapter.java
@@ -1,5 +1,8 @@
-package eu.wswieciejutra;
+package eu.wswieciejutra.persistence;
 
+import eu.wswieciejutra.Branch;
+import eu.wswieciejutra.BranchInterface;
+import eu.wswieciejutra.Repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
