+++ monolith/src/main/java/eu/wswieciejutra/config/OpenAPIConfig.java
@@ -1,4 +1,4 @@
-package eu.wswieciejutra;
+package eu.wswieciejutra.config;
 
 import io.swagger.v3.oas.models.OpenAPI;
 import io.swagger.v3.oas.models.info.Contact;
